move libraries to libs
