+++ instrumentino/libs/abs_ratio.py
